requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
openai>=1.0.0
pytest>=9.0.0
